move typetree logic into a new file
